% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dist}
\alias{dist}
\alias{sqDist}
\title{Distance calculation}
\usage{
dist(a1, a2)

sqDist(a1, a2)
}
\arguments{
\item{a1,a2}{individuals}
}
\description{
dist calculates the distance between two individuals, and should be used by the end user for
exploring simulation results.
}
\details{
sqDist calculates the squared distance between two individuals.
For most application (such as determining neighbors) knowing the square distance is enough,
and square roots are computationally expensive. Also, this function does not check for obvious errors
(such as the parameters not being individuals, or being individuals from different landscapes),
as it is designed to be as efficient as possible.
}
\examples{
L <- Landscape(bound.condition="p", numb.cells = 50) # x ranges from -25 to 25
S <- Species(L)
a1 <- Individual(S, 25, 25)
a2 <- Individual(S, -20, -20)
sqDist(a1, a2)
}

