% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individual.R
\name{Individual}
\alias{Individual}
\title{Creates a new individual for the simulation}
\usage{
Individual(landscape, species, x = 0, y = 0)
}
\arguments{
\item{landscape}{in which \code{\link{Landscape}} this individual lives}

\item{species}{from which \code{\link{Species}} this individual is}

\item{x,y}{initial position}
}
\description{
Creates a new individual for the simulation
}

