% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-accessors.R
\name{population}
\alias{popSize}
\alias{population}
\title{Accessors methods for class Landscape}
\usage{
population(scape)

popSize(scape)
}
\arguments{
\item{scape}{The landscape}
}
\description{
Use these functions to access the status of a landscape or its enclosed population
}

