% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landscape.R
\name{Landscape}
\alias{Landscape}
\title{Landscape generators}
\usage{
Landscape(numb.cells = 100, cell.size = 1, land.shape = c("square",
  "circle"), type = c("random", "blob", "fahrig"),
  bound.condition = c("absorptive", "periodical", "reflexive"), cover = 1,
  frag = NULL)
}
\arguments{
\item{numb.cells}{integer. Represents both the lenght AND width of the landscape, so numb.cells=100 creates a 100x100 landscape}

\item{cell.size}{numerical. How many meters/kilometers should a single cell represent?}

\item{land.shape,bound.condition}{character. Control whether the shape of the land should be treated as a square (default) or a circle inscribed in this square; and what is the boundary condition. These parameters have no effect on the actual landscape generated, but will be used passed to the simulation.}

\item{type}{character. Currently, three types of habitat fragmentation routines are implemented: random fragmentation (noise), contiguous blob removal (random walk removal) and Fahrig fragmentation.}

\item{cover}{numerical, between 0 and 1. What fraction of the landscape should consist on habitat?}

\item{frag}{recuperation parameter for the Fahrig method. Is ignored for other types.}
}
\description{
This function generates a landscape object, which represent a fragmented environment in which
the simulations will run.
}

