% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Species.R
\name{populate}
\alias{populate}
\title{Populates a species}
\usage{
populate(species, N, FUN = NULL, ...)
}
\arguments{
\item{species}{the species}

\item{N}{number of individuals to be generated}

\item{FUN}{the generating function for initial positions. Will be called once for x and once for y
for each individual. Common choices are NULL (for starting position zero), "runif" or "rnorm"}

\item{\dots}{further arguments for FUN. If FUN is a random number generator such as "runif", "n" might be ommited}
}
\description{
Creates N individuals for the specified species, on starting positions generated by function FUN.
}
\section{Details}{

The individual creation applies the landscape boundary condition to each individual. So, for the case of
absorptive boundaries, the individual may be created and instantly killed, as it was created outside the
boundaries of the landscape. Thus, it is advisable to check the population size after calling this function.
}
\examples{
L <- Landscape(numb.cells=10)
S <- Species(L)
populate(S, 20, "runif", min=-5, max=5)
}

