% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{runSSim}
\alias{Exponential}
\alias{Logistic}
\alias{runSSim}
\title{Simulation routines}
\usage{
runSSim(maxtime, N, ...)

Exponential(maxtime, N, ...)

Logistic(maxtime, N, ...)
}
\arguments{
\item{maxtime}{Maximum simulation running time}

\item{N}{initial population size}

\item{\dots}{further arguments that will be passed to Landscape or Species. All options specified
on \code{\dots} and not used for these functions will be passed to \code{\link{populate}}}
}
\description{
Set up a new \code{\link{Landscape}} with preset Species and Individiduals, and run a Gillespie
algorithm until a final time is reached. \code{runSSim} runs a single-species simulation and passes
all parameters as named to the Landscape or Species constructor.
}
\details{
Exponential and Logistic are just wrappers for runSSim that enforce that the model is a simple
exponential / logistic growth.
}

