% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape.R
\name{Landscape}
\alias{Landscape}
\alias{plot.landscape}
\title{Landscape generators}
\usage{
Landscape(numb.cells = 100, type = c("random", "blob", "fahrig", "disc"),
  bound.condition = c("absorptive", "periodical", "reflexive"), cover = 1,
  frag = NULL)

\method{plot}{landscape}(x, col1 = "darkgreen", col2 = "grey70", ...)
}
\arguments{
\item{numb.cells}{integer. Represents both the lenght AND width of the landscape, so numb.cells=100 creates a 100x100 landscape}

\item{type}{character. Currently, three types of habitat fragmentation routines are implemented: random fragmentation (noise), contiguous blob removal (random walk removal) and Fahrig fragmentation. Also, you can specify "disc" to create a landscape that is close to a circular disc; in this case, the disc is fully made of habitat and the parameter cover is ignored.}

\item{bound.condition}{character. Control what is the boundary condition. This parameter have no effect on the actual landscape generated, but will be passed to the simulation.}

\item{cover}{numerical, between 0 and 1. What fraction of the landscape should consist on habitat?}

\item{frag}{fragmentation parameter for the Fahrig method. Is ignored for other types. Higher values
are close to random noise, while lower values result in more realistic landshapes.}

\item{col1}{habitat color}

\item{col2}{non-habitat color}
}
\description{
This function generates a landscape object, which represent a fragmented environment in which
the simulations will run.
}
\examples{
plot(Landscape(cover=0.7, type="f", frag=0.01))
}

