% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Species.R
\name{Species}
\alias{Species}
\title{Creates a new species}
\usage{
Species(landscape, birth.rate = 1, death.rate = 0.1, incl.birth = 500,
  incl.death = 0, radius = 1, matrix.death = 10, move.rate = 1,
  step = 1, visual.angle = base::pi/4)
}
\arguments{
\item{landscape}{the \code{\link{Landscape}} in which this species lives}

\item{birth.rate}{the intrinsic birth rate of the species}

\item{death.rate}{the intrinsic death rate of the species}

\item{incl.birth}{the slope in which the birth rate decreases with increasing density}

\item{incl.death}{the slope in which the death rate increases with increasing density}

\item{radius}{the perception radius of the species.}

\item{matrix.death}{a constant multiplying the death rate in non-habitat patches (Currently unimplemented!)}

\item{move.rate}{the intrinsic movement rate of the species}

\item{step}{the length of the step for each movement}

\item{visual.angle}{currently, controls how much the movement orientation changes between steps}
}
\description{
Currently, every simulation contains a single species that must be created using this function.
}
\section{WARNING}{


Every time the function Species is called, a new species is created
and pushed to the landscape's species list.
}

